+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Guardian screen pre-release](https://github.com/multiversx/mx-sdk-dapp/pull/739)]
 - [Fixed transactions data and signature parsing](https://github.com/multiversx/mx-sdk-dapp/pull/738)]
 
 ## [[v2.11.3]](https://github.com/multiversx/mx-sdk-dapp/pull/737)] - 2023-04-21